refactor(render): split responsive table projection
